requirements.txt
scipy==1.8.0
numpy==1.19.2
Pillow==9.0.1